+++ lib/semian/adapter.rb
@@ -9,19 +9,25 @@ module Semian
     end
 
     def semian_resource
-      @semian_resource ||= case semian_options
+      return @semian_resource if @semian_resource
+
+      case semian_options
       when false
-        UnprotectedResource.new(semian_identifier)
+        @semian_resource = UnprotectedResource.new(semian_identifier)
       when nil
         Semian.logger.info("Semian is not configured for #{self.class.name}: #{semian_identifier}")
-        UnprotectedResource.new(semian_identifier)
+        @semian_resource = UnprotectedResource.new(semian_identifier)
       else
         options = semian_options.dup
         options.delete(:name)
         options[:consumer] = self
         options[:exceptions] ||= []
         options[:exceptions] += resource_exceptions
-        ::Semian.retrieve_or_register(semian_identifier, **options)
+        resource = ::Semian.retrieve_or_register(semian_identifier, **options)
+
+        @semian_resource = resource unless options.fetch(:dynamic, false)
+
+        resource
       end
     end
 
@@ -53,7 +59,11 @@ module Semian
       return @semian_options if defined? @semian_options
 
       options = raw_semian_options
-      @semian_options = options && options.map { |k, v| [k.to_sym, v] }.to_h
+
+      symbolized_options = options && options.transform_keys(&:to_sym) # rubocop:disable Style/SafeNavigation
+      symbolized_options.tap do
+        @semian_options = symbolized_options if !symbolized_options || !symbolized_options.fetch(:dynamic, false)
+      end
     end
 
     def raw_semian_options