+++ test/helpers/adapter_helper.rb
@@ -19,6 +19,28 @@ module Semian
     end
   end
 
+  module DynamicAdapterTest
+    include Semian::Adapter
+
+    def semian_identifier
+      :dynamic_semian_adapter_test
+    end
+
+    def raw_semian_options
+      {
+        bulkhead: false,
+        error_threshold: 1,
+        error_timeout: 1,
+        dynamic: true,
+        success_threshold: @current_success_threshold += 1,
+      }
+    end
+
+    def resource_exceptions
+      []
+    end
+  end
+
   class AdapterTestClient
     include AdapterTest
 
@@ -28,6 +50,18 @@ module Semian
         error_threshold: 1,
         error_timeout: 1,
       )
+    end
+
+    def ==(other)
+      inspect == other.inspect
+    end
+  end
+
+  class DynamicAdapterTestClient
+    include DynamicAdapterTest
+
+    def initialize(**args)
+      @current_success_threshold = 1
     end
 
     def ==(other)