+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Redis Readonly errors won't trigger semian open circuit. (#489)
+* New `dynamic` options, allowing a per-request resource configuration with the HTTP adapter. (#485)
 
 # v0.18.1